Introduce torch.sym_add, variadic add
Partially addresses https://github.com/pytorch/pytorch/issues/128150

When you have big sums of values, we end up computing long chains of
binary addition in our FX graph representation.  Not only is this ugly,
it also is quadratic, as the sympy.Add constructor is O(N) in number
of arguments.  Instead, ensure that we maintain the summation as a
single FX node so we can do the entire addition all in one go.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: 4cc81b86521c6693b91eb835b1f8b0fe8a73ad62
Pull Request resolved: https://github.com/pytorch/pytorch/pull/136429